+++ src-ui/app/views/dashboard-view/views/status-automations-view/tabs/status-automations-general-tab/status-automations-general-tab.component.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-status-automations-general-tab',
+  templateUrl: './status-automations-general-tab.component.html',
+  styleUrls: ['./status-automations-general-tab.component.scss']
+})
+export class StatusAutomationsGeneralTabComponent {
+
+}